requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
websockets>=13.0
elevenlabs==1.0.0
pydantic>=2.7.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
supabase>=2.3.0
asyncpg>=0.29.0
